deeply embedded router .push es are weird but okay

+++ src/components/useWorkspaceFileMgmt.tsx
@@ -3,6 +3,7 @@ import { TreeNode } from "@/clientdb/filetree";
 import { Workspace } from "@/clientdb/Workspace";
 import { useFileTreeMenuContext } from "@/components/FileTreeContext";
 import { WorkspaceRouteType } from "@/context";
+import { NotFoundError } from "@/lib/errors";
 import { AbsPath, absPath, isAncestor, reduceLineage, relPath, RelPath } from "@/lib/paths";
 import { usePathname, useRouter } from "next/navigation";
 
@@ -15,9 +16,9 @@ export function useWorkspaceFileMgmt(currentWorkspace: Workspace, workspaceRoute
   const renameFile = async (oldNode: TreeNode, newFullPath: AbsPath) => {
     const { path } = await currentWorkspace.renameFile(oldNode, newFullPath);
 
-    if (workspaceRoute.path?.str === oldNode.path.str) {
-      router.push(currentWorkspace.replaceUrlPath(pathname, oldNode.path, path));
-    }
+    // if (workspaceRoute.path?.str === oldNode.path.str) {
+    //   router.push(currentWorkspace.replaceUrlPath(pathname, oldNode.path, path));
+    // }
     return path;
   };
 
@@ -29,9 +30,9 @@ export function useWorkspaceFileMgmt(currentWorkspace: Workspace, workspaceRoute
   };
   const newFile = async (path: AbsPath, content = "") => {
     const newPath = await currentWorkspace.newFile(path.dirname(), path.basename(), content);
-    // if (!workspaceRoute.path) {
-    //   router.push(currentWorkspace.resolveFileUrl(newPath));
-    // }
+    if (!workspaceRoute.path) {
+      router.push(currentWorkspace.resolveFileUrl(newPath));
+    }
     return newPath;
   };
   const newDir = async (path: AbsPath) => {
@@ -45,34 +46,40 @@ export function useWorkspaceFileMgmt(currentWorkspace: Workspace, workspaceRoute
     }
     if (!range.length) return;
 
-    // if (workspaceRoute.path && range.includes(workspaceRoute.path.str)) {
-    //   const firstFile = currentWorkspace.disk.getFirstFile();
-    //   if (firstFile) {
-    //     router.push(currentWorkspace.resolveFileUrl(firstFile.path));
-    //   } else {
-    //     router.push(currentWorkspace.href);
-    //   }
-    // }
-
     const paths = reduceLineage(range).map((pathStr) => absPath(pathStr));
-    // console.log(selectedRange, paths);
-    return Promise.all(paths.map((path) => currentWorkspace.removeFile(path)));
+    try {
+      await Promise.all(paths.map((path) => currentWorkspace.removeFile(path)));
+      if (workspaceRoute.path && range.includes(workspaceRoute.path.str)) {
+        router.push(currentWorkspace.tryFirstFileUrl());
+      }
+    } catch (e) {
+      if (e instanceof NotFoundError) {
+        console.error(e); //???
+      } else {
+        throw e;
+      }
+    }
+    setFocused(null);
   };
 
   const removeFocusedFile = async () => {
+    if (!focused || !currentWorkspace.disk.pathExists(focused)) return;
     const focusedNode = currentWorkspace.nodeFromPath(focused);
     if (!focusedNode) return;
     if (focusedNode.path.str === "/") return;
 
-    await currentWorkspace.removeFile(focusedNode.path);
-    // if (workspaceRoute.path?.str === focusedNode.path.str) {
-    //   const firstFile = currentWorkspace.disk.getFirstFile();
-    //   if (firstFile) {
-    //     router.push(currentWorkspace.resolveFileUrl(firstFile.path));
-    //   } else {
-    //     router.push(currentWorkspace.href);
-    //   }
-    // }
+    try {
+      await currentWorkspace.removeFile(focusedNode.path);
+    } catch (e) {
+      if (e instanceof NotFoundError) {
+        //do nothing its okay
+      } else {
+        throw e;
+      }
+    }
+    if (workspaceRoute.path?.str === focusedNode.path.str) {
+      router.push(currentWorkspace.tryFirstFileUrl());
+    }
   };
   const cancelNew = () => {
     setVirtual(null);
